requirements.txt
asgiref==3.11.0
Django==6.0.1
django-ckeditor==6.7.3
django-filter==25.2
django-js-asset==3.1.2
django-markdownx==4.0.9
Markdown==3.10.1
pillow==12.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
whitenoise==6.11.0
